move struct to mgmt types
